2-TV-Shows-API-REST/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
pyspark==3.5.5